requirements.txt
numpy
pandas
opencv-python
py-cpuinfo
WMI

